requirements.txt
pandas==2.2.2
requests==2.32.3
tiktoken=0.7.0
matplotlib==3.9.1
openai==1.35.13
dvc==3.51.2
dvc-gdrive==3.0.1
bs4==0.0.2
python-dotenv==1.0.1
numpy==2.0.0
scikit-learn==1.5.0
xgboost==2.1.0
nltk==3.8.1
spacy==3.7.5
gradio==4.37.2
wandb==0.17.4
joblib==1.4.2
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
opentsne==1.0.1
umap-learn==0.5.6
codecarbon==2.5.0
seaborn==0.13.2
